refactor: lib -> libs
