+++ src/components/ui/calendar.tsx
@@ -8,7 +8,7 @@ import {
 } from "lucide-react";
 import { DayButton, DayPicker, getDefaultClassNames } from "react-day-picker";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 import { Button, buttonVariants } from "@/components/ui/button";
 
 function Calendar({
