+++ src/components/ui/resizable.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import { GripVerticalIcon } from "lucide-react";
 import * as ResizablePrimitive from "react-resizable-panels";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 function ResizablePanelGroup({
   className,
