+++ src/components/ui/scroll-area.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as ScrollAreaPrimitive from "@radix-ui/react-scroll-area";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 function ScrollArea({
   className,
