+++ src/components/ui/table.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/libs/utils";
 
 function Table({ className, ...props }: React.ComponentProps<"table">) {
   return (
